Build RPM packages
