+++ linux/rpm/config/jdk-10-hotspot/provides.txt
@@ -1,0 +1,8 @@
+java
+java-10
+java-10-openjdk
+java-openjdk
+jre
+jre-10
+jre-10-openjdk
+jre-openjdk
