+++ linux/rpm/config/jdk-11-openj9/provides.txt
@@ -1,0 +1,8 @@
+java
+java-11
+java-11-openjdk
+java-openjdk
+jre
+jre-11
+jre-11-openjdk
+jre-openjdk
