+++ linux/rpm/config/jdk-12-openj9/provides.txt
@@ -1,0 +1,8 @@
+java
+java-12
+java-12-openjdk
+java-openjdk
+jre
+jre-12
+jre-12-openjdk
+jre-openjdk
