+++ linux/rpm/config/jdk-8-openj9/provides.txt
@@ -1,0 +1,8 @@
+java
+java-1.8.0
+java-1.8.0-openjdk
+java-openjdk
+jre
+jre-1.8.0
+jre-1.8.0-openjdk
+jre-openjdk
