+++ linux/rpm/config/jdk-9-hotspot/dependencies.txt
@@ -1,0 +1,17 @@
+/bin/sh
+/usr/sbin/alternatives
+ca-certificates
+libX11.so.6
+libXext.so.6
+libXi.so.6
+libXrender.so.1
+libXtst.so.6
+libasound.so.2
+libc.so.6
+libelf.so.1
+libdl.so.2
+libgcc_s.so.1
+libm.so.6
+libpthread.so.0
+libthread_db.so.1
+libz.so.1
