+++ linux/rpm/config/jdk-9-openj9/provides.txt
@@ -1,0 +1,8 @@
+java
+java-9
+java-9-openjdk
+java-openjdk
+jre
+jre-9
+jre-9-openjdk
+jre-openjdk
